BL-6165: mot_test_rfr_map_comment assigned as int in TAPI result set

+++ api/src/main/java/uk/gov/dvsa/mot/persist/jdbc/MotTestReadDaoJdbc.java
@@ -498,7 +498,7 @@ public class MotTestReadDaoJdbc implements MotTestReadDao {
         ResultSetRowMapper<MotTestRfrMapCustomDescription> mapper = rs -> {
             MotTestRfrMapCustomDescription motTestRfrMapCustomDescription = new MotTestRfrMapCustomDescription();
 
-            motTestRfrMapCustomDescription.setId(rs.getInt(1));
+            motTestRfrMapCustomDescription.setId(rs.getLong(1));
             motTestRfrMapCustomDescription.setMotTestCurrentRfrMap(parent);
             motTestRfrMapCustomDescription.setCustomDescription(rs.getString(2));
             motTestRfrMapCustomDescription.setCreatedBy(rs.getInt(3));
